+++ Sources/Contour.Transport.RabbitMq/Internal/ByteArrayPayloadConverter.cs
@@ -1,0 +1,21 @@
+using System;
+
+using Contour.Serialization;
+
+namespace Contour.Transport.RabbitMq.Internal
+{
+    internal class ByteArrayPayloadConverter : IPayloadConverter
+    {
+        public string ContentType { get; }
+
+        public byte[] FromObject(object payload)
+        {
+            return (byte[])payload;
+        }
+
+        public object ToObject(byte[] payload, Type targetType)
+        {
+            return (object)payload;
+        }
+    }
+}