+++ Sources/Contour.Transport.RabbitMq/Internal/Listener.cs
@@ -13,6 +13,7 @@ using Contour.Helpers;
 using Contour.Helpers.Timing;
 using Contour.Receiving;
 using Contour.Receiving.Consumers;
+using Contour.Serialization;
 using Contour.Validation;
 
 using Contour.Transport.RabbitMq.Internal;
@@ -64,7 +65,7 @@ namespace Contour.Transport.RabbitMq.Internal
         private readonly IBusContext busContext;
         private readonly IRabbitConnection connection;
         
-        private readonly ConcurrentBag<RabbitChannel> channels = new ConcurrentBag<RabbitChannel>();
+        private readonly ConcurrentBag<IRabbitChannel> channels = new ConcurrentBag<IRabbitChannel>();
         private CancellationTokenSource cancellationTokenSource;
 
         /// <summary>
@@ -161,9 +162,27 @@ namespace Contour.Transport.RabbitMq.Internal
         /// <returns>
         /// Входящее сообщение.
         /// </returns>
-        public RabbitDelivery BuildDeliveryFrom(RabbitChannel deliveryChannel, BasicDeliverEventArgs args)
+        public RabbitDelivery BuildDeliveryFrom(IRabbitChannel deliveryChannel, BasicDeliverEventArgs args)
         {
-            return new RabbitDelivery(this.busContext, deliveryChannel, args, this.ReceiverOptions.IsAcceptRequired());
+            IPayloadConverterResolver payloadConverterResolver = this.ReceiverOptions.GetPayloadConverterResolver();
+            IPayloadConverter payloadConverter;
+            try
+            {
+                payloadConverter = payloadConverterResolver.ResolveConverter(args.BasicProperties.ContentType);
+            }
+            catch (Exception ex)
+            {
+                this.OnFailure(
+                    new RabbitDelivery(
+                        this.busContext, 
+                        deliveryChannel, args, 
+                        this.ReceiverOptions.IsAcceptRequired(), 
+                        new ByteArrayPayloadConverter()),
+                    ex);
+                throw;
+            }
+            return new RabbitDelivery(this.busContext, deliveryChannel, args, this.ReceiverOptions.IsAcceptRequired(), payloadConverter);
+
         }
 
         /// <summary>
@@ -368,7 +387,7 @@ namespace Contour.Transport.RabbitMq.Internal
                     }
                 }
 
-                RabbitChannel channel;
+                IRabbitChannel channel;
                 while (this.channels.TryTake(out channel))
                 {
                     try
@@ -422,7 +441,7 @@ namespace Contour.Transport.RabbitMq.Internal
         {
             try
             {
-                RabbitChannel channel;
+                IRabbitChannel channel;
                 var consumer = this.InitializeConsumer(token, out channel);
 
                 while (!token.IsCancellationRequested)
@@ -465,7 +484,7 @@ namespace Contour.Transport.RabbitMq.Internal
             return new Expectation(d => this.BuildResponse(d, expectedResponseType), timeoutTicket);
         }
         
-        private CancellableQueueingConsumer InitializeConsumer(CancellationToken token, out RabbitChannel channel)
+        private CancellableQueueingConsumer InitializeConsumer(CancellationToken token, out IRabbitChannel channel)
         {
             // Opening a new channel may lead to a new connection creation
             channel = this.connection.OpenChannel(token);