+++ Sources/Contour/Configuration/BusConfiguration.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Configuration;
 using System.Linq;
 
@@ -29,6 +30,8 @@ namespace Contour.Configuration
         /// The _filters.
         /// </summary>
         private readonly IList<IMessageExchangeFilter> filters = new List<IMessageExchangeFilter>();
+
+        private readonly IList<IPayloadConverter> converters = new List<IPayloadConverter>();
 
         /// <summary>
         /// The _message label resolver.
@@ -149,10 +152,7 @@ namespace Contour.Configuration
         /// </summary>
         public SenderOptions SenderDefaults { get; internal set; }
 
-        /// <summary>
-        /// Gets the serializer.
-        /// </summary>
-        public IPayloadConverter Serializer { get; private set; }
+        public ReadOnlyCollection<IPayloadConverter> Converters => new ReadOnlyCollection<IPayloadConverter>(this.converters);
 
         /// <summary>
         /// Gets the validator registry.
@@ -513,7 +513,7 @@ namespace Contour.Configuration
         /// </param>
         public void UsePayloadConverter(IPayloadConverter converter)
         {
-            this.Serializer = converter;
+            this.converters.Add(converter);
         }
 
         /// <summary>
@@ -558,9 +558,9 @@ namespace Contour.Configuration
         {
             Logger.Trace(m => m("Validating. Connection string - [{0}], endpoint name - [{1}], incoming labels - [{2}], outgoing labels - [{3}]", this.EndpointOptions.GetConnectionString().HasValue ? this.EndpointOptions.GetConnectionString().Value : "N/A", this.Endpoint, this.ReceiverConfigurations != null ? string.Join(";", this.ReceiverConfigurations.Select(x => x.Label)) : "null", this.SenderConfigurations != null ? string.Join(";", this.SenderConfigurations.Select(x => x.Label)) : "null"));
 
-            if (this.Serializer == null)
+            if (!this.Converters.Any())
             {
-                throw new BusConfigurationException("PayloadConverter is not set.");
+                throw new BusConfigurationException("Payload converter is not set. UsePayloadConverter should be called before transport configuration.");
             }
 
             if (this.BusFactoryFunc == null)